src/backend/requirements-dev.txt
fastapi
mlflow-skinny
pandas
scikit-learn
python-dotenv
uvicorn

# Testing
pytest
black
flake8
httpx